Fix #25 auto load issue
The composer auto loader was being included via `include_once`. This was
preventing the bootstrap file from executing when the composer autoloader
had already been `include`d by PHPUnit. The composer auto loader appears
to be designed to be included many times, as the `getLoader()` function
caches it's results.

This patch fixes the issue.